Rename Body to Container
