+++ src/components/Body/Content.js
@@ -1,0 +1,7 @@
+import React from "react";
+import { ContentContainer } from "../../pages/styles";
+const Body = ({ children }) => {
+  return <ContentContainer>{children}</ContentContainer>;
+};
+
+export default Body;
