+++ src/components/index.js
@@ -4,5 +4,5 @@ export { default as SideBar } from "./Body/SideBar";
 export { default as Task } from "./Body/Task";
 export { default as TaskInput } from "../components/Body/TaskInput";
 export { default as Title } from "./Header/Title";
-export { default as Body } from "./Body/Body";
+export { default as Body } from "./Body/Content";
 export { default as Header } from "./Header/Header";
