+++ src/pages/Themes.js
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import SettingComponent from "../components/main/SettingComponent";
 
-const SettingBodyContainer = styled.div`
+const SettingContentContainer = styled.div`
   display: flex;
   flex-direction: column;
   height: 100%;
@@ -19,7 +19,7 @@ const SettingTitleContainer = styled.div`
   align-items: center;
 `;
 
-const SettingContentContainer = styled.div`
+const SettingBodyContainer = styled.div`
   height: 85%;
   width: 100%;
   display: flex;
@@ -31,10 +31,10 @@ const SettingContentContainer = styled.div`
 const Themes = () => {
   return (
     <SettingComponent>
-      <SettingBodyContainer>
+      <SettingContentContainer>
         <SettingTitleContainer>Theme settings</SettingTitleContainer>
-        <SettingContentContainer>Coming Soon</SettingContentContainer>
-      </SettingBodyContainer>
+        <SettingBodyContainer>Coming Soon</SettingBodyContainer>
+      </SettingContentContainer>
     </SettingComponent>
   );
 };
